Refactors the propagation interfaces, removing the Propagator make the whole component more cohesive.
