+++ composer.json
@@ -23,7 +23,7 @@
         "psr-4": {
             "OpenTracing\\": "./src/OpenTracing/"
         },
-        "files": ["./src/OpenTracing/Ext/Tags.php"]
+        "files": ["./src/OpenTracing/Ext/Tags.php", "./src/OpenTracing/Propagation/Formats.php"]
     },
     "autoload-dev": {
         "psr-4": {
